+++ docs/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [Unreleased]
+
+### Added
+
+- `/pbdb` help command to show bot information.
+
+---
+
 ## [0.6.0]: 2023-07-13
 
 ### Added